Make sure that expected values were not influenced by earlier tests
In some tests the status of elements is changed (hide -> show).
We need to be careful that before we test a function, this change
is controlled. So when an element should change from hide to show
make sure that the element is in the state hide and not already shown.